refactor: moving code
